# perf: faster pattern matcher