--- a/tinygrad/codegen/uops.py
+++ b/tinygrad/codegen/uops.py
@@ -110,22 +110,43 @@ def __unmatch(m1:Union[T, Set[T]], m2:T) -> bool:
   return False
 
 def _match(uop:UOp, pat:UPat, store:Dict[str, UOp]) -> bool:
-  if pat.name in store and store[pat.name] is not uop: return False
+  if pat.arg is not None and __unmatch(pat.arg, uop.arg):return False, store
+  if pat.op is not None and __unmatch(pat.op, uop.op):return False, store
+  if pat.name in store and store[pat.name] is not uop: return False, store
   if pat.name is not None: store[pat.name] = uop
-  if pat.arg is not None and __unmatch(pat.arg, uop.arg): return False
-  if pat.dtype is not None and uop.dtype is not None and __unmatch(pat.dtype, uop.dtype): return False
-  if pat.op is not None and __unmatch(pat.op, uop.op): return False
-  if pat.src is None: return True
+  if pat.dtype is not None and uop.dtype is not None and __unmatch(pat.dtype, uop.dtype): return False, store
+  if pat.src is None:return True, store
   # only one if it's a tuple
   # try all permutations if it's a list
   # repeat if it's a UPat
   for vp in itertools.permutations(pat.src) if isinstance(pat.src,list) else ([pat.src] if isinstance(pat.src,tuple) else [(pat.src,)*len(uop.src)]):
-    if len(uop.src) != len(vp) and (len(uop.src) not in pat.allow_len): return False
+    if len(uop.src) != len(vp) and (len(uop.src) not in pat.allow_len): return False, store
     new_store = store.copy()
-    if all(_match(uu, vv, new_store) for uu, vv in zip(uop.src, vp)):
+    for uu, vv in zip(uop.src, vp):
+      res, new_store = _match(uu, vv, new_store)
+      if not res: break
+    else:
       store.update(new_store)
-      return True
-  return False
+      return True, store
+  return False, store
+
+@functools.lru_cache(None)
+def recursive_rewrite(pm: PatternMatcher, up:UOp) -> UOp:
+  recurse_cnt = 0
+  while (rewritten := rewrite_on_match(pm, up)) is not None:
+    assert recurse_cnt < 100, f"recursive_rewrite looped {up} <--> {rewritten}"
+    up = rewritten
+    recurse_cnt += 1
+  if up.src and (new_src := tuple(recursive_rewrite(pm, x) for x in up.src)) != up.src:
+      return UOp(up.op, up.dtype, new_src, up.arg)
+  return up
+
+@functools.lru_cache(None)
+def rewrite_on_match(pm, uop:UOp) -> Optional[UOp]:
+  for p,fxn in itertools.chain(pm.pdict[(uop.op, uop.arg)], pm.pdict[(uop.op, None)]):
+    res, store = _match(uop, p, {})
+    if res: return fxn(**store)
+  return None
 
 class PatternMatcher:
   def __init__(self, patterns:List[Tuple[Union[UPat, UOp], Callable]]):
@@ -139,26 +160,10 @@ class PatternMatcher:
         for uop in p.op: self.pdict[(uop, p.arg)].append((p, fxn))
       else:
         self.pdict[(p.op, p.arg)].append((p, fxn))
+    
 
-  @functools.lru_cache(None)
-  def recursive_rewrite(self, u:UOp) -> UOp:
-    recurse_cnt = 0
-    up = u
-    # locally recursively rewrite
-    while (rewritten := self.rewrite_on_match(up)) is not None:
-      assert recurse_cnt < 100, f"recursive_rewrite looped {up} <--> {rewritten}"
-      up = rewritten
-    if up.src:
-      new_src = tuple(self.recursive_rewrite(x) for x in up.src)
-      if new_src != up.src:
-        up = UOp(up.op, up.dtype, new_src, up.arg)
-    return up
 
-  def rewrite_on_match(self, uop:UOp) -> Optional[UOp]:
-    for p,fxn in itertools.chain(self.pdict[(uop.op, uop.arg)], self.pdict[(uop.op, None)]):
-      store: Dict[str, UOp] = {}
-      if _match(uop, p, store): return fxn(**store)
-    return None
+
 
 def sum_collapse(phi_input, loop, val1, val2):
   for v1,v2 in [(val1, val2), (val2, val1)]:
@@ -305,10 +310,9 @@ class UOpGraph:
       print(f"{i:4d} {str(u.op):20s}: {str(u.dtype) if u.dtype is not None else '':25s} " f"{str([self.uops.index(x) for x in u.src]):32s} {u.arg}")
 
   def graph_rewrite(self, sink, pm):
-    # recursive rewrite
     changed = getenv("UOPS_REWRITE", True)
     while changed:
-      rewritten = pm.recursive_rewrite(sink)
+      rewritten = recursive_rewrite(pm, sink)
       changed = sink != rewritten
       sink = rewritten
     self.nodes[sink] = sink