722.45/530.45 = 1.36

Timings with `PROFILE=0 python -O test/external/external_benchmark_schedule.py` (best out of 5)

master
```
***** model forward in  15.73 ms
***** model schedule in   5.30 ms
***** model lower in 722.45 ms
```

This branch
```
***** model forward in  15.29 ms
***** model schedule in   5.06 ms
***** model lower in 530.45 ms
```

* rewrite methods on patternmatcher, cache recursive rewrite

cc: @DKormann Any ideas on this? Frozen UOp should allow for fast hashes and better caching.