@DKormann 2024-06-24
> Frozen UOp should allow for fast hashes and better caching.